pyproject.toml
[project]
name = "amazon-order-mcp"
version = "0.1.0"
description = "MCP server wrapper for amazon-orders library"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "amazon-orders>=4.0.0",
    "python-dotenv>=1.0.0",
    "nest-asyncio>=1.6.0",
]

[project.scripts]
amazon-order-mcp = "amazon_order_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
